solve that sum of two integers and string input and output problems by using 1D array, for loop, if,else statement, and cin,cout function

+++ c++_programming/chapter01/SimpleAdder.cpp
@@ -1,7 +1,10 @@
 # include <iostream>
 using namespace std;
 int main(){
-    int v1,v2,r;
+    // 입력이 정수형일 때
+    // int v1,v2,r;
+    // 입력이 실수형일 때
+    double v1,v2,r;
     cout<<"첫 번째 숫자입력: ";
     cin>>v1;
     cout<<"두 번째 숫자입력: ";
